Can trigger off of k and kdot (curvature and derivative of curvature) of the curvature of the head.
This allows us to have the laser shine only when the head is swinging towards the ventral side and has swung a certain distance that way.

As of now you can specify to trigger when k> kthresh and when sign(kdot)==desiredSign.

This allows you to select ranges of curvature in different quadrants.

Think of k and kdot as defining a phase space, as one would find for simple harmonic motion.