Minimum illumination and refractory period illumination timing feature works.
This commit addresses two issues:

1) When the illumination had timed out, but the worm still had curvature that indicated that the illumination needed to be on, before this commit the software would turn off the DLP transiently and then turn it back on. By adding a line in the HandleCurvaturePhaseAnalysis() function to explicitly turn on the DLP this was resolved.

2)The curvature information that the HandleCurvaturePhaseAnalysis was getting was a frame old. I moved in down in the sequence of events in main.cpp so that we find the segmentation first, so that we can then do our curvature analysis based on this.

As a corrolary to 1 and 2, I moved the ClearDLPifNotDisplayingNow down. This is somewhat unfortunate, because for the time it takes too do segmentation and curvature phase analysis, the DLP could be needlessly on. But i justify it because we won't know if we awant to turn off the DLP in the case of curvature triggering until after we do the segmentation. At most I estimate we will be exposing the worm to 2 or 3ms of extra illumination.