BUGFIX: added 50ms delay to initialization of second thread
After merging in the the smooth and erode function, I believe what is going on is that the addition of two more sliderbars makes the SetupGUI function, run by the main thread, too slow to get ready in time, so that by the time the second thread is looking to update, the variables don't exist. (Or something like that must be going on).

In any event, after merging the following commit:

1febb7c42c9882bae37d

everythign was crashing. So but adding in this short 50ms delay fixed all that.