Stage Active Zone. Committed. Untested.
For the DualMag setup we'd like the stage velocity to be exquisitely sensitive to when the worm makes small local movements near the center, but we'd like to cap the maximum stage speed, so that if the worm is somehow way far off, we don't zip the stage at some ridiculous speed.

Therefore the velocity of the stage as a function of the worms distance from the center is a sharp v with flat wingtips that extend to the end of the image.