Stripped to essentials and recapitulate linker problem
Even running makeBFtest.mak -f testFG gets a linker error that complains
that the references to the Ci functions are undefined.

This is with mingw32. There's almost nothing here. Its hellow world, the
talk2FrameGrabber.cpp adn teh BitFlow SDK. Note a downgraded to the
BitFlow 5.2 SDK just to control for variables.