!!!CRITICAL BUGFIX!!! Protocol caused errant illumination
Previously, using a protocol for illumination was apparently causing the
software to send garbage illumination patterns superimposed onto the
legitimate illumination patterns to the DLP. The garbage patterns mostly
consisted of a mirrored and scaeld version of the DLP illumination as well
as the outline of the worm.

The garbage patterns became mingled with the legitimate illumination
pattern when I created a new image using cvCreateImage in
IllumWormPorotocol.c line 846.  Somehow this image was not initialized to
zero, and instead it seemed to contain other recently used images.

Now I have zeroed this image by adding in a call to cvSetZero.

So to summarize, for the past I-don't-know-how-long I have been sendign
garbage to the DLP superimposed on top of the illumination pattern
whenever I illuminate using a protocol. I will now need to go back into
the code and find out how long this has been a problem. I seem to
recall rewriting a section of that code so I will try to find when that
was.