Fixed Illumination Out of Bounds of DMD: Intermittant
This is an intermittant and rare bug that has plagued the software since its
inception. However it is extremely rare.

An initial analysis of data from all experiments conducted between October
23rd and Nov 4 of 2013 suggest that less then 1/10th of 1% of all frames
when the DMD was on were affected.

The bug has to do with the softwares behavior when the illumination
pattern of the worm is out of bounds of the mirrors on the DMD. The DMD
covers only a subset of the region of the field of view. So it is possible
for the worm to be in the field of view of the camera, but inaccessible to
the DMD.

In the past, when this happened, the software would try to generate
illumination polygons, but it would map vertices of the polygon that fell
out of bounds of the DMD to the mirror's DMD located at 0,0. When the
software then filled in the polygon defined by these vertices, it would
mostly be the correct polygon, but it would have a vertex at the corner of
the DMD so there might be a triangle of light that extended across the DMD
towards the corner.

This only happened when the illumination pattern for the worm left the
bounds of the DMD. By design, this is a very rare occurance. The feedback
from the stage works hard to keep the worm centered in the middle of the
field of view (and as a corallary, in the middle of the DMD). But if the
worm undergoes a quick acceleration while we are illuminating, for
example, it might just happen that this errror could occur.

Note we only discovered this bug while doing some un-physiological
experiments where we fed in images of the worm to the software, instead of
a real worm.

Nonetheless, I no have fixed this unexpected behavior. Now when the worm's
illumination pattern walks outside of the field of view, the user is
warned. Moreover, there is no errant illuminattion pattern. The portion of
the illumination pattern that is out of bounds is obviously not
illuminated, but everything else remains as expected.

To do this I modified the MATLAB routine that generates the LookUp table
based on the paired calibration points.

I also modifed the C code to provide a more helpful error.